+++ src/Loggers/PaymentLogger.php
@@ -2,6 +2,7 @@
 
 namespace PlacetoPay\Loggers;
 
+use Exception;
 use \FileLogger;
 use \PrestaShopLogger;
 
@@ -33,14 +34,21 @@ class PaymentLogger
         $line = null
     ) {
         $format = sprintf("[%s:%d] => [%d]\n %s", $file, $line, $errorCode, $message);
+        $logSuccess = true;
 
-        self::getLogInstance()->log($format, $severity, $errorCode);
+        try {
+            self::getLogInstance()->log($format, $severity, $errorCode);
+        } catch (Exception $exception) {
+            $logSuccess = false;
 
-        if ($severity >= self::WARNING) {
-            self::logInDatabase($message, $severity, $errorCode);
+            self::logInDatabase($exception->getMessage(), self::WARNING, $exception->getCode());
         }
 
-        return true;
+        if ($severity >= self::WARNING || !$logSuccess) {
+            return self::logInDatabase($message, $severity, $errorCode);
+        }
+
+        return $logSuccess;
     }
 
     /**
@@ -51,7 +59,13 @@ class PaymentLogger
      */
     public static function logInDatabase($message, $severity = self::INFO, $errorCode = null)
     {
-        PrestaShopLogger::addLog($message, $severity, $errorCode, getModuleName(), $errorCode > 0 ? $errorCode : 999);
+        try {
+            $errorCode = $errorCode > 0 ? $errorCode : 999;
+
+            PrestaShopLogger::addLog($message, $severity, $errorCode, getModuleName(), $errorCode);
+        } catch (Exception $exception) {
+            return false;
+        }
 
         return true;
     }
@@ -61,7 +75,7 @@ class PaymentLogger
      */
     public static function getLogFilename()
     {
-        $logfile = sprintf('%s_%s.log', getModuleName(), date('Y-m-d'));
+        $logfile = sprintf('%s_%s_%s.log', (isDebugEnable() ? 'dev' : 'prod'), @date('Ymd'), getModuleName());
 
         // PS < 1.7.0.0
         $pathLogs = '/log/';