+++ src/Models/PaymentPrestaShop.php
@@ -512,7 +512,7 @@ class PaymentPrestaShop extends PaymentModule
                 $formErrors[] = sprintf('%s %s', $this->ll('Payment description'), $this->ll('is required.'));
             }
 
-            // Configuration Connection
+            // Connection Configuration
             if (!Tools::getValue(self::EXPIRATION_TIME_MINUTES)) {
                 $formErrors[] = sprintf('%s %s', $this->ll('Expiration time to pay'), $this->ll('is required.'));
             } elseif (filter_var(Tools::getValue(self::EXPIRATION_TIME_MINUTES), FILTER_VALIDATE_INT) === false
@@ -584,7 +584,7 @@ class PaymentPrestaShop extends PaymentModule
                 Configuration::updateValue(self::STOCK_REINJECT, Tools::getValue(self::STOCK_REINJECT));
             }
 
-            // Configuration Connection
+            // Connection Configuration
             Configuration::updateValue(self::COUNTRY, Tools::getValue(self::COUNTRY));
             Configuration::updateValue(self::ENVIRONMENT, Tools::getValue(self::ENVIRONMENT));
             // Set or clean custom URL
@@ -611,10 +611,12 @@ class PaymentPrestaShop extends PaymentModule
     {
         $this->smarty->assign(
             [
+                'is_set_credentials' => $this->isSetCredentials(),
                 'version' => $this->getPluginVersion(),
                 'url_notification' => $this->getUrl('process.php'),
                 'schedule_task' => $this->getPathScheduleTask(),
-                'is_set_credentials' => $this->isSetCredentials(),
+                'log_file' => $this->getPathLogFile(),
+                'log_database' => $this->context->link->getAdminLink('AdminLogs'),
             ]
         );
 
@@ -858,7 +860,7 @@ class PaymentPrestaShop extends PaymentModule
         $fieldsFormConnection = [
             'form' => [
                 'legend' => [
-                    'title' => $this->ll('Configuration Connection'),
+                    'title' => $this->ll('Connection Configuration'),
                     'icon' => 'icon-rocket'
                 ],
                 'input' => [
@@ -1432,6 +1434,14 @@ class PaymentPrestaShop extends PaymentModule
     private function getPathScheduleTask()
     {
         return fixPath($this->getPathThisModule() . '/sonda.php');
+    }
+
+    /**
+     * @return string
+     */
+    private function getPathLogFile()
+    {
+        return PaymentLogger::getLogFilename();
     }
 
     /**