+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Added
-- Add dev-* commands in Makefile
-
 ### Updated
 - Update dependencies guzzle/guzzle from 5.3.2 => 5.3.3
 - Update README file with [mailtrap.io](https://mailtrap.io/)
+- Update format name logfile, it is: \[dev|prod\]_YYYYMMDD_placetopayment.log
+- Update commands in Makefile
+
+### Fixed
+- Fix error when log in file or database failed
 
 ## [3.4.0] - 2018-07-25